+++ example/plugins/keyboard.py
@@ -47,29 +47,29 @@ async def _(msg, ctx):
     await ctx.reply("Message with keyboard", keyboard=KEYBOARD_STRING)
 
 
-@plugin.vk.on_payload([1])
+@plugin.vk.on_payloads([1])
 async def _(msg, ctx):
     await ctx.reply("Your choice was 1 (from handler)")
 
 
-@plugin.vk.on_payload([2])
+@plugin.vk.on_payloads([2])
 async def _(msg, ctx):
     await ctx.reply("Your choice was 2 (from handler)")
 
 
-@plugin.vk.on_payload([3])
+@plugin.vk.on_payloads([3])
 async def _(msg, ctx):
     await ctx.reply("Your choice was 3 (from handler)")
 
 
 # This will never be called because of handler below
-@plugin.vk.on_payload([4])
+@plugin.vk.on_payloads([4])
 async def _(msg, ctx):
     await ctx.reply("Your choice was 4 (from handler)")
 
 
 # Intercept all the messages with payloads and check their payload
-@plugin.on_any_message(priority=10)
+@plugin.on_messages(priority=10)
 async def _(msg, ctx):
     # Only messages from vkontakte can have payload
     if ctx.backend.get_identity() != "vkontakte":